chore: move index file
